Split RakNet backend implementations (#2)
